Changes: -Replace some annotations in controller -Validations added in service -New method added in RestRepository -Couse class added

+++ src/main/java/com/Service/StudentService.java
@@ -6,8 +6,13 @@ import com.Entity.Student;
 import com.Utilities.StringUtilities;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.http.HttpRequest;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
+import org.springframework.web.client.HttpServerErrorException;
 
+import java.net.HttpRetryException;
 import java.util.Optional;
 
 @Service
@@ -22,28 +27,27 @@ public class StudentService {
     }
 
     public Optional<Student> getStudentById(int id) {
-        return this.studentDao.findById(id);
+         Optional<Student> stu = this.studentDao.findById(id);
+         if (stu.isPresent()) return stu;
+         else throw new HttpClientErrorException(HttpStatus.BAD_REQUEST, "Estudiante inexistente");
     }
 
     public void removeStudentById(int id) {
-        // aca puede ir logica de que si encuentra el estudiante lo devuelva sino devuelve un http error o algo asi
-        this.studentDao.deleteById(id);
+        if (studentDao.findById(id).isPresent()) this.studentDao.deleteById(id);
+        else throw new HttpClientErrorException(HttpStatus.BAD_REQUEST, "Estudiante inexistente");
     }
 
     public void insertAndUpdateStudent(Student student) {
-        Optional<Student> stu = studentDao.findById(student.getId());
+        Optional<Student> stu = studentDao.findStudentByName(student.getName());
         if (stu.isPresent()) {
             if (student.studentAttrAreCorrect()) {
                 stu.get().setName(student.getName());
                 stu.get().setCourse(student.getCourse());
                 this.studentDao.save(stu.get());
-            } else {
-
-            }
+            } else throw new HttpClientErrorException(HttpStatus.BAD_REQUEST, "Falta Materia");
         } else {
-            if (student.studentAttrAreCorrect()) {
-                this.studentDao.save(student);
-            }
+            if (student.studentAttrAreCorrect()) this.studentDao.save(student);
+            else throw new HttpClientErrorException(HttpStatus.BAD_REQUEST, "Falta Nombre o Materia");
         }
     }
